+++ src/Telegram.Bot/Types/Enums/FileType.cs
@@ -1,8 +1,12 @@
-﻿namespace Telegram.Bot.Types.Enums
+﻿using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+
+namespace Telegram.Bot.Types.Enums
 {
     /// <summary>
     /// Type of a <see cref="IInputFile"/>
     /// </summary>
+    [JsonConverter(typeof(StringEnumConverter), true)]
     public enum FileType
     {
         /// <summary>