+++ src/Telegram.Bot/Types/Enums/MessageEntityType.cs
@@ -7,19 +7,17 @@ namespace Telegram.Bot.Types.Enums
     /// <summary>
     /// Type of a <see cref="MessageEntity"/>
     /// </summary>
-    [JsonConverter(typeof(StringEnumConverter))]
+    [JsonConverter(typeof(StringEnumConverter), true)]
     public enum MessageEntityType
     {
         /// <summary>
         /// A mentioned <see cref="User"/>
         /// </summary>
-        [EnumMember(Value = "mention")]
         Mention,
 
         /// <summary>
         /// A searchable Hashtag
         /// </summary>
-        [EnumMember(Value = "hashtag")]
         Hashtag,
 
         /// <summary>
@@ -31,37 +29,31 @@ namespace Telegram.Bot.Types.Enums
         /// <summary>
         /// An url
         /// </summary>
-        [EnumMember(Value = "url")]
         Url,
 
         /// <summary>
         /// An email
         /// </summary>
-        [EnumMember(Value = "email")]
         Email,
 
         /// <summary>
         /// Bold text
         /// </summary>
-        [EnumMember(Value = "bold")]
         Bold,
 
         /// <summary>
         /// Italic text
         /// </summary>
-        [EnumMember(Value = "italic")]
         Italic,
 
         /// <summary>
         /// Monowidth string
         /// </summary>
-        [EnumMember(Value = "code")]
         Code,
 
         /// <summary>
         /// Monowidth block
         /// </summary>
-        [EnumMember(Value = "pre")]
         Pre,
 
         /// <summary>