+++ test/Telegram.Bot.Tests.Integ/Admin Bot/ChatMemberAdministrationTests.cs
@@ -85,8 +85,8 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
                 .GetUpdatesAsync(u =>
                     u.Message.Chat.Type == ChatType.Supergroup &&
                     u.Message.Chat.Id.ToString() == _fixture.SupergroupChat.Id.ToString() &&
-                    u.Message.Type == MessageType.ServiceMessage,
-                updateTypes: UpdateType.MessageUpdate)
+                    u.Message.Type == MessageType.Service,
+                updateTypes: UpdateType.Message)
             ).Single();
 
             await _fixture.UpdateReceiver.DiscardNewUpdatesAsync();